These changes bring squishit back inline to the fork quad/graphics made back with version 0.6.
-Added in a default debug predicate to check for a querry variable
called debugmode.
-Added the ability to retrieve a cacheed instance of a bundle from
a generated file name that has a hash.
-Added a prefix to the path locations to default our corporate frameworks
script path. (Could maybe use BaseOutputHref for this?)
-Fixed a problem with having a '#' in the path of the project.  The app
would replace the '#' in the path and file name with the hash value where
it should just do the file name.  The current fix is will need to be
revisited because I do not think it is enough. (just looks for the last
'#' and replaces it.)
-Added back the IBundle interface since this is what our corporate
framework used.  This is an attempt to cause less problems upgrading.
-Also added back a couple of bundle methods that were renamed with
obsolete tags.